+++ src/types/QuickPickItem.ts
@@ -1,0 +1,4 @@
+import * as vscode from 'vscode';
+export interface TodoQuickPickItem extends vscode.QuickPickItem {
+    line: number;
+}